Fix decode TCP flags with u32 (#968)
* Fix decode TCP flags with u32

Fixes #946

* new uint conversion func

* simplify test

+++ pkg/utils/tcp_flags.go
@@ -1,7 +1,7 @@
 package utils
 
 type tcpFlag struct {
-	value uint16
+	value uint
 	name  string
 }
 
@@ -19,7 +19,7 @@ var tcpFlags = []tcpFlag{
 	{value: 1024, name: "RST_ACK"},
 }
 
-func DecodeTCPFlags(bitfield uint16) []string {
+func DecodeTCPFlags(bitfield uint) []string {
 	var values []string
 	for _, flag := range tcpFlags {
 		if bitfield&flag.value != 0 {
